Activity-class

Represents the provenance of a Synapse Entity.

name: name of the Activity

description: a short text description of the Activity

used: Either a list of: - `reference objects <http://docs.synapse.org/rest/org/sagebionetworks/repo/model/Reference.html>`_ (e.g. ``['targetId':'syn123456', 'targetVersionNumber':1]``) - a list of Synapse Entities or Entity IDs - a list of URL's

executed: A code resource that was executed to generate the Entity.

data: A dictionary representation of an Activity, with fields 'name', 'description' and 'used' (a list of reference objects)

See also: `Provenance in Synapse <https://sagebionetworks.jira.com/wiki/display/PLFM/Analysis+Provenance+in+Synapse>`_ and the `W3C's provenance ontology <http://www.w3.org/TR/prov-o/>`_

Format

An R6 class object.

Methods

Activity(name=NULL, description=NULL, used=NULL, executed=NULL, data=list()): Represents the provenance of a Synapse Entity.

:param name: name of the Activity :param description: a short text description of the Activity :param used: Either a list of: - `reference objects <http://docs.synapse.org/rest/org/sagebionetworks/repo/model/Reference.html>`_ (e.g. ``['targetId':'syn123456', 'targetVersionNumber':1]``) - a list of Synapse Entities or Entity IDs - a list of URL's :param executed: A code resource that was executed to generate the Entity. :param data: A dictionary representation of an Activity, with fields 'name', 'description' and 'used' (a list of reference objects)

See also: `Provenance in Synapse <https://sagebionetworks.jira.com/wiki/display/PLFM/Analysis+Provenance+in+Synapse>`_ and the `W3C's provenance ontology <http://www.w3.org/TR/prov-o/>`_

executed(target=NULL, targetVersion=NULL, url=NULL, name=NULL): Add a code resource that was executed during the activity. See :py:func:`synapseclient.activity.Activity.used`

used(target=NULL, targetVersion=NULL, wasExecuted=NULL, url=NULL, name=NULL): Add a resource used by the activity.

This method tries to be as permissive as possible. It accepts a string which might be a synapse ID or a URL, a synapse entity, a UsedEntity or UsedURL dictionary or a list containing any combination of these.

In addition, named parameters can be used to specify the fields of either a UsedEntity or a UsedURL. If target and optionally targetVersion are specified, create a UsedEntity. If url and optionally name are specified, create a UsedURL.

It is an error to specify both target/targetVersion parameters and url/name parameters in the same call. To add multiple UsedEntities and UsedURLs, make a separate call for each or pass in a list.

In case of conflicting settings for wasExecuted both inside an object and with a parameter, the parameter wins. For example, this UsedURL will have wasExecuted set to False::

activity.used('url':'http://google.com', 'name':'Goog', 'wasExecuted':True, wasExecuted=False)

Entity examples::

activity.used('syn12345') activity.used(entity) activity.used(target=entity, targetVersion=2) activity.used(codeEntity, wasExecuted=True) activity.used('reference':'target':'syn12345', 'targetVersion':1, 'wasExecuted':False)

URL examples::

activity.used('http://mydomain.com/my/awesome/data.RData') activity.used(url='http://mydomain.com/my/awesome/data.RData', name='Awesome Data') activity.used(url='https://github.com/joe_hacker/code_repo', name='Gnarly hacks', wasExecuted=True) activity.used('url':'https://github.com/joe_hacker/code_repo', 'name':'Gnarly hacks', wasExecuted=True)

List example::

activity.used(['syn12345', 'syn23456', entity, 'reference':'target':'syn100009', 'targetVersion':2, 'wasExecuted':True, 'http://mydomain.com/my/awesome/data.RData'])

usedEntity(target, targetVersion=NULL, wasExecuted=FALSE): **Deprecated**

See :py:func:`synapseclient.Activity.used`.

usedURL(url, name=NULL, wasExecuted=FALSE): **Deprecated**

See :py:func:`synapseclient.Activity.used`.