File-class

Represents a file in Synapse.

When a File object is stored, the associated local file or its URL will be stored in Synapse. A File must have a path (or URL) and a parent.

path: Location to be represented by this File

name: Name of the file in Synapse, not to be confused with the name within the path

parent: Project or Folder where this File is stored

synapseStore: Whether the File should be uploaded or if only the path should be stored. Defaults to True (file should be uploaded)

contentType: Manually specify Content-type header, for example "application/png" or "application/json; charset=UTF-8"

dataFileHandleId: Defining an existing dataFileHandleId will use the existing dataFileHandleId The creator of the file must also be the owner of the dataFileHandleId to have permission to store the file

Format

An R6 class object.

Methods

File(path=NULL, parent=NULL, synapseStore=TRUE, properties=NULL, annotations=NULL, local_state=NULL, ...): Represents a file in Synapse.

When a File object is stored, the associated local file or its URL will be stored in Synapse. A File must have a path (or URL) and a parent.

:param path: Location to be represented by this File :param name: Name of the file in Synapse, not to be confused with the name within the path :param parent: Project or Folder where this File is stored :param synapseStore: Whether the File should be uploaded or if only the path should be stored. Defaults to True (file should be uploaded) :param contentType: Manually specify Content-type header, for example "application/png" or "application/json; charset=UTF-8" :param dataFileHandleId: Defining an existing dataFileHandleId will use the existing dataFileHandleId The creator of the file must also be the owner of the dataFileHandleId to have permission to store the file ::

data = File('/path/to/file/data.xyz', parent=folder) data = syn.store(data)

deleteURI(versionNumber=NULL):

has_key(key): Is the given key a property or annotation?

keys(): Returns a set of property and annotation keys

local_state(state=NULL): Set or get the object's internal state, excluding properties, or annotations.

:param state: A dictionary

postURI():

putURI():