StoreTableOptions

Options for storing Table-like entities.

Usage

StoreTableOptions(dry_run=FALSE, job_timeout=600L)

Format

An S3-style class object.

Arguments

dry_run

(bool) If TRUE, will not actually store but will log what would be done.

job_timeout

(int) Maximum time to wait for table schema update job to complete.

Value

NULL

Methods

StoreTableOptions(dry_run=FALSE, job_timeout=600L): Constructor for StoreTableOptions