Table : reorder_column

Reorder a column in the table. Note that this does not store the column in Synapse. You must call the synStore() function on this Table instance to store the column in Synapse. This is a convenience function to eliminate the need to manually reorder the named list stored in the $columns attribute.

You must ensure that the index is within the bounds of the number of columns in the table. If you pass in an index that is out of bounds the column will be added to the end of the list.

Usage

synReorderColumn(instance, name, index)

Arguments

instance

(Table) The Table instance to operate on.

name

(str) The name of the column to reorder.

index

(int) The index to move the column to starting with 0.

Value

NULL

Examples

if (FALSE) { # \dontrun{
## Reordering a column
    # This example shows how you may reorder a column in a table and then store
    # the change back in Synapse.
    library(synapser)
    synLogin()

    table = Table(id = "syn1234") |> synGet(table_options = TableOptions(include_columns = TRUE))

    # Move the column to the beginning of the list
    table |> synReorderColumn(name = "my_column", index = 0L)
    table |> synStore()
} # }