Wiki-class

Represents a wiki page in Synapse with content specified in markdown.

title: Title of the Wiki

owner: Parent Entity that the Wiki will belong to

markdown: Content of the Wiki (cannot be defined if markdownFile is defined)

markdownFile: Path to file which contains the Content of Wiki (cannot be defined if markdown is defined)

attachments: List of paths to files to attach

fileHandles: List of file handle IDs representing files to be attached

parentWikiId: (optional) For subpages, specify parent wiki page

Format

An R6 class object.

Methods

Wiki(...): Represents a wiki page in Synapse with content specified in markdown.

:param title: Title of the Wiki :param owner: Parent Entity that the Wiki will belong to :param markdown: Content of the Wiki (cannot be defined if markdownFile is defined) :param markdownFile: Path to file which contains the Content of Wiki (cannot be defined if markdown is defined) :param attachments: List of paths to files to attach :param fileHandles: List of file handle IDs representing files to be attached :param parentWikiId: (optional) For subpages, specify parent wiki page

deleteURI(): For internal use.

getURI(): For internal use.

json(): Returns the JSON representation of the Wiki object.

postURI(): For internal use.

putURI(): For internal use.

update_markdown(markdown=NULL, markdown_file=NULL): Updates the wiki's markdown. Specify only one of markdown and markdown_file :param markdown: text that will become the markdown :param markdown_file: path to a file. Its contents will be the markdown