synGet

Gets a Synapse entity from the repository service.

synGet(entity, version=NULL, downloadFile=NULL, downloadLocation=NULL, followLink=NULL, ifcollision=NULL, limitSearch=NULL)

Arguments

entity
A Synapse ID, a Synapse Entity object, a named list in which 'id' maps to a Synapse ID or a local file that is stored in Synapse (found by hash of file)

version
optional named parameter: The specific version to get. Defaults to the most recent version.

downloadFile
optional named parameter: Whether associated files(s) should be downloaded. Defaults to TRUE

downloadLocation
optional named parameter: Directory where to download the Synapse File Entity. Defaults to the local cache.

followLink
optional named parameter: Whether the link returns the target Entity. Defaults to FALSE

ifcollision
optional named parameter: Determines how to handle file collisions. May be "overwrite.local", "keep.local", or "keep.both". Defaults to "keep.both".

limitSearch
optional named parameter: a Synanpse ID used to limit the search in Synapse if entity is specified as a local file. That is, if the file is stored in multiple locations in Synapse only the ones in the specified folder/project will be returned.

Details

Value

A new Synapse Entity object of the appropriate type

References

Note

See also

Examples

# NOT RUN {
## download file
file <- synGet('syn1906479')
print(file$properties$name)
print(file$path)
# }