synLogin

Logs the user in.

If no login arguments are provided or only username is provided, login() will attempt to log in using information from these sources (in order of preference):

synapseConfig file (in user home folder unless configured otherwise)

User defined arguments during a CLI session

User's Personal Access Token (aka: Synapse Auth Token) from the environment variable: SYNAPSE_AUTH_TOKEN

Retrieves user's authentication token from AWS SSM Parameter store (if configured)

synLogin(email=NULL, silent=FALSE, authToken=NULL, cache_client=TRUE)

Arguments

email

Synapse user name (or an email address associated with a Synapse account). Defaults to NULL.

silent

Suppresses the "Welcome ...!" message. Defaults to FALSE.

authToken

A bearer authorization token, e.g. a personal access token, can be used in lieu of a
password or apiKey

cache_client

Whether to cache the Synapse client object in the Synapse module. Defaults to TRUE. When set to TRUE anywhere a Synapse object is optional you do not need to pass an instance of Synapse to that function, method, or class.

rememberMe

DEPRECATED. Whether the authentication information should be cached in your operating system's credential storage.
**GNOME Keyring** (recommended) or **KWallet** is recommonded to be installed for credential storage on **Linux** systems.
If it is not installed/setup, credentials will be stored as PLAIN-TEXT file with read and write permissions for the current user only (chmod 600).
On Windows and Mac OS, a default credentials storage exists so it will be preferred over the plain-text file.

password

DEPRECATED. Please use authToken (Synapse personal access token)

apiKey

DEPRECATED. Base64 encoded Synapse API key

sessionToken

DEPRECATED. User's current session token. Using this field will ignore the following fields: email, password, apiKey

forced

DEPRECATED. Defaults to FALSE. Bypass the credential cache if set.

Examples

if (FALSE) { # \dontrun{

# with a token, e.g. an access obtained from your Synapse profile
synLogin(authToken=token)

} # }