synLogin

Logs the user in.

synLogin(email=NULL, password=NULL, apiKey=NULL, sessionToken=NULL, rememberMe=FALSE, silent=FALSE, forced=FALSE)

Arguments

email

Synapse user name (or an email address associated with a Synapse account)

password

**!!WILL BE DEPRECATED!!** password. Please use authToken (Synapse personal access token)

apiKey

**!!WILL BE DEPRECATED!!** Base64 encoded Synapse API key

sessionToken

**!!DEPRECATED FIELD!!** User's current session token. Using this field will ignore the following fields: email, password, apiKey

rememberMe

Whether the authentication information should be cached in your operating system's credential storage.
**GNOME Keyring** (recommended) or **KWallet** is recommonded to be installed for credential storage on **Linux** systems.
If it is not installed/setup, credentials will be stored as PLAIN-TEXT file with read and write permissions for the current user only (chmod 600).
On Windows and Mac OS, a default credentials storage exists so it will be preferred over the plain-text file.

silent

Defaults to FALSE. Suppresses the "Welcome ...!" message.

forced

Defaults to FALSE. Bypass the credential cache if set.

authToken

A bearer authorization token, e.g. a personal access token, can be used in lieu of a
password or apiKey

Details

Valid combinations of login() arguments:

- email/username and password (**WILL BE DEPRECATED**)
- email/username and apiKey (Base64 encoded string) (**WILL BE DEPRECATED**)
- authToken
- sessionToken (**DEPRECATED**)
If no login arguments are provided or only username is provided, login() will attempt to log in using information from these sources (in order of preference):
- User's personal access token from environment the variable: SYNAPSE_AUTH_TOKEN
- .synapseConfig file (in user home folder unless configured otherwise)
- cached credentials from previous `login()` where `rememberMe=True` was passed as a parameter

Examples

if (FALSE) {

# with username/pass
synLogin('myUsername', 'secretPassword')

# with a token, e.g. an access obtained from your Synapse profile
synLogin(authToken=token)

}