synLogin

Logs the user in.

synLogin(email=NULL, password=NULL, apiKey=NULL, sessionToken=NULL, rememberMe=FALSE, silent=FALSE, forced=FALSE)

Arguments

email
Synapse user name (or an email address associated with a Synapse account)

password
password

apiKey
Base64 encoded Synapse API key

sessionToken
A previously obtained session token

rememberMe
Whether the authentication information should be cached locally for usage across sessions and clients.

silent
Defaults to FALSE. Suppresses the "Welcome ...!" message.

forced
Defaults to FALSE. Bypass the credential cache if set.

Details

Authenticates the user using the given credentials (in order of preference): - supplied synapse user name (or email) and password - supplied email and API key (base 64 encoded) - supplied session token - supplied email and cached API key - most recent cached email and API key - email in the configuration file and cached API key - email and API key in the configuration file - email and password in the configuraton file - session token in the configuration file

Value

References

Note

See also

Examples

# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }
# NOT RUN {

# }